considerar que o user_o pode ser o user_1

+++ FastAPI/crud/user_relations.py
@@ -87,7 +87,7 @@ def update_user_relations(db: Session, user_relations: UserRelationsPost, user_i
 
 def delete_user_relations(db: Session, user_id_0: int, user_id_1: int):
     try:
-        db.query(UserRelations).filter_by(user_0=user_id_0, user_1=user_id_1).delete()
+        db.query(UserRelations).filter_by(or_(and_(user_0=user_id_0, user_1=user_id_1), and_(user_0=user_id_1, user_1=user_id_0))).delete()
         db.commit()
         status = True
     except Exception:
